made vote patch api request generic to handle from within vote component, and set navigation to new article after composition, with changes so vote and delete components conditionally render correctly

+++ src/components/Delete.jsx
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Delete = props => {
-    const { deleteItem, commentId } = props
+    const { deleteItem, commentId, user_id, author_id, newArticle } = props
+    if (!newArticle && user_id !== author_id) return null
     return (
         <div>
             <button onClick={()=> deleteItem(commentId)}>Delete</button>
